dx11shaderanalyse / rename scripts: Include signatures in bytecode hash
After thinking about what sections should be hashed and looking at some
of the shaders in MGSVTPP it occurred to me that ignoring the signature
sections in the hash might be a mistake, as it is valid for two separate
shaders with different signatures to compile to the same bytecode, but
may not be valid to replace them with another shader with a different
signature (if nothing else the debug layer is sure to complain).

To this end, modify the bytecode hash to include all possible input,
output and patch constant signature sections (as of SM5.1), and update
the scripts to use this new hash.

+++ rename2embedded.sh
@@ -13,8 +13,14 @@ for shader_txt in *.txt; do
 	shader_bin=$(echo "$shader_txt" | sed 's/\(................-.s\).*/\1.bin/')
 	suffix=$(echo "$shader_txt" | sed 's/................\(-.s.*\)/\1/')
 
-	new_hash=$(dx11shaderanalyse.py -v "../ShaderCache/$shader_bin" | \
-		awk '/hash/ {print $3}' | grep . | sed 's/^\(................\).*/\1/')
+	orig_bin="../ShaderCache/$shader_bin"
+	if [ ! -e "$orig_bin" ]; then
+		echo "$orig_bin" not found
+		continue
+	fi
+
+	new_hash=$(dx11shaderanalyse.py -v "$orig_bin" | \
+		awk '/Embedded hash/ {print $3}' | sed 's/^\(................\).*/\1/')
 
 	mv -v "$shader_txt" "$new_hash$suffix"
 done
